feat: Implement shop registration feature with address handling and unique slug generation

+++ backend/src/main/java/com/dev/backend/modules/shop/dto/RegisterShopRequest.java
@@ -25,8 +25,8 @@ public record RegisterShopRequest(
         String bankNumber,
         String ownerName,
 
-        Long provinceId,
-        Long districtId,
+        Integer provinceId,
+        Integer districtId,
         String wardCode,
         String street
 ) {}
